+++ app/routes/app+/rewards.index.tsx
@@ -1,44 +1,28 @@
-import { Link, useSubmit } from "@remix-run/react";
-import { useRef } from "react";
-import { Checkbox } from "~/components/checkbox";
-import { Pagination } from "~/components/pagination/pagination";
-import { Modal } from "~/components/modal";
-import { ValidatedInput } from "~/components/input";
-import { Button } from "~/components/button";
-import { useState } from "react";
-import { ValidatedCombobox } from "~/components/combobox";
-import { withZod } from "@remix-validated-form/with-zod";
-import { ValidatedForm } from "remix-validated-form";
-import { Container } from "~/components/container";
-import RewardsTab from "~/features/rewards-tab";
-import { Card } from "~/components/card";
-import { fromNow } from "~/utils/date";
-import { Header, Table, Row } from "~/components/table";
-import { CheckCircleIcon, MagnifyingGlassIcon } from "@heroicons/react/20/solid";
-import { getUser } from "~/services/session.server";
-import { findAllWalletsForUser } from "~/services/wallet.server";
+import { MagnifyingGlassIcon } from "@heroicons/react/20/solid";
+import type { Token, Wallet } from "@prisma/client";
+import { useSubmit } from "@remix-run/react";
 import type { DataFunctionArgs } from "@remix-run/server-runtime";
+import { withZod } from "@remix-validated-form/with-zod";
+import { useRef } from "react";
+import { getParamsOrFail } from "remix-params-helper";
 import { typedjson, useTypedLoaderData } from "remix-typedjson";
-import { createBlockchainTransactionStateMachine } from "~/utils/machine";
-import type { ClaimRewardContractData } from "~/hooks/use-claim-reward";
-import { useMachine } from "@xstate/react";
-import { ClaimRewardWeb3Button } from "~/features/web3-button/claim-reward";
+import { ValidatedForm } from "remix-validated-form";
 import invariant from "tiny-invariant";
-import type { SendTransactionResult } from "@wagmi/core";
-import { defaultNotifyTransactionActions } from "~/features/web3-transaction-toasts";
-import { searchUserSubmissions } from "~/services/submissions.server";
+import { Field, Label, ValidatedSelect } from "~/components";
+import { Checkbox } from "~/components/checkbox";
+import { ValidatedCombobox } from "~/components/combobox";
+import { Container } from "~/components/container";
+import { ValidatedInput } from "~/components/input";
+import { Pagination } from "~/components/pagination/pagination";
 import type { SubmissionWithServiceRequest } from "~/domain/submission";
 import { RewardsSearchSchema } from "~/domain/submission";
-import { Field, Label, ValidatedSelect } from "~/components";
+import { RewardsCards } from "~/features/my-rewards/rewards-card-mobile";
+import { RewardsTable } from "~/features/my-rewards/rewards-table-desktop";
+import RewardsTab from "~/features/rewards-tab";
+import { getUser } from "~/services/session.server";
+import { searchUserSubmissions } from "~/services/submissions.server";
 import { listTokens } from "~/services/tokens.server";
-import type { Token, Wallet } from "@prisma/client";
-import { getParamsOrFail } from "remix-params-helper";
-import { fromTokenAmount, toNetworkName, toTokenAbbreviation } from "~/utils/helpers";
-import type { EthersError } from "~/features/web3-button/types";
-import { RPCError } from "~/features/rpc-error";
-import { useGetReward } from "~/hooks/use-get-reward";
-import { useHasPerformed } from "~/hooks/use-has-performed";
-import { RewardBadge } from "~/components/reward-badge";
+import { findAllWalletsForUser } from "~/services/wallet.server";
 
 const validator = withZod(RewardsSearchSchema);
 
@@ -118,326 +102,6 @@ function RewardsListView({
       <div className="block lg:hidden">
         <RewardsCards rewards={rewards} wallets={wallets} tokens={tokens} />
       </div>
-    </>
-  );
-}
-
-function RewardsTable({
-  rewards,
-  wallets,
-  tokens,
-}: {
-  rewards: SubmissionWithServiceRequest[];
-  wallets: Wallet[];
-  tokens: Token[];
-}) {
-  return (
-    <Table>
-      <Header columns={12} className="mb-2">
-        <Header.Column span={3}>Challenge Title</Header.Column>
-        <Header.Column span={3}>Reward</Header.Column>
-        <Header.Column span={2}>Submitted</Header.Column>
-        <Header.Column span={3}>Rewarded</Header.Column>
-        <Header.Column>Status</Header.Column>
-      </Header>
-      {rewards.map((r) => {
-        return (
-          <RewardsTableRow
-            key={`${r.id}${r.serviceRequestId}${r.laborMarketAddress}`}
-            reward={r}
-            wallets={wallets}
-            tokens={tokens}
-          />
-        );
-      })}
-    </Table>
-  );
-}
-
-function RewardsTableRow({
-  reward,
-  wallets,
-  tokens,
-}: {
-  reward: SubmissionWithServiceRequest;
-  wallets: Wallet[];
-  tokens: Token[];
-}) {
-  const contractReward = useGetReward({
-    laborMarketAddress: reward.laborMarketAddress as `0x${string}`,
-    submissionId: reward.id,
-  });
-  const hasClaimed = useHasPerformed({
-    laborMarketAddress: reward.laborMarketAddress as `0x${string}`,
-    id: reward.id,
-    action: "HAS_CLAIMED",
-  });
-  const token = tokens.find((t) => t.contractAddress === reward.sr.configuration.pToken);
-  const showReward = contractReward !== undefined && hasClaimed === false;
-  const showRewarded = contractReward !== undefined && hasClaimed === true;
-
-  return (
-    <Row columns={12}>
-      <Row.Column span={3}>
-        <p>{reward.sr.appData?.title}</p>
-      </Row.Column>
-      <Row.Column span={3}>
-        {showReward ? (
-          <RewardBadge
-            amount={fromTokenAmount(contractReward[0].toString())}
-            token={token?.symbol ?? "Unknown Token"}
-            rMETRIC={contractReward[1].toNumber()}
-          />
-        ) : (
-          <span>--</span>
-        )}
-      </Row.Column>
-      <Row.Column span={2} className="text-black">
-        {fromNow(reward.createdAtBlockTimestamp)}{" "}
-      </Row.Column>
-      <Row.Column span={3} className="text-black" color="dark.3">
-        {showRewarded ? (
-          <RewardBadge
-            amount={fromTokenAmount(contractReward[0].toString())}
-            token={token?.symbol ?? "Unknown Token"}
-            rMETRIC={contractReward[1].toNumber()}
-          />
-        ) : (
-          <span>--</span>
-        )}
-      </Row.Column>
-      <Row.Column>
-        {hasClaimed === false ? (
-          <ClaimButton reward={reward} wallets={wallets} tokens={tokens} />
-        ) : hasClaimed === true ? (
-          <span>Claimed</span>
-        ) : (
-          <></>
-        )}
-      </Row.Column>
-    </Row>
-  );
-}
-
-function RewardsCards({
-  rewards,
-  wallets,
-  tokens,
-}: {
-  rewards: SubmissionWithServiceRequest[];
-  wallets: Wallet[];
-  tokens: Token[];
-}) {
-  return (
-    <div className="space-y-4">
-      {rewards.map((r) => {
-        return <RewardCard key={`${r.laborMarketAddress}_${r.id}`} reward={r} wallets={wallets} tokens={tokens} />;
-      })}
-    </div>
-  );
-}
-
-function RewardCard({
-  reward,
-  wallets,
-  tokens,
-}: {
-  reward: SubmissionWithServiceRequest;
-  wallets: Wallet[];
-  tokens: Token[];
-}) {
-  const contractReward = useGetReward({
-    laborMarketAddress: reward.laborMarketAddress as `0x${string}`,
-    submissionId: reward.id,
-  });
-  const hasClaimed = useHasPerformed({
-    laborMarketAddress: reward.laborMarketAddress as `0x${string}`,
-    id: reward.id,
-    action: "HAS_CLAIMED",
-  });
-  const token = tokens.find((t) => t.contractAddress === reward.sr.configuration.pToken);
-  const showReward = contractReward !== undefined && hasClaimed === false;
-  const showRewarded = contractReward !== undefined && hasClaimed === true;
-  return (
-    <Card className="grid grid-cols-2 gap-y-3 gap-x-1 items-center px-2 py-5">
-      <div>Challenge Title</div>
-      <p>{reward.sr.appData?.title}</p>
-      <div>Reward</div>
-      <div>
-        {showReward ? (
-          <RewardBadge
-            amount={fromTokenAmount(contractReward[0].toString())}
-            token={token?.symbol ?? "Unknown Token"}
-            rMETRIC={contractReward[1].toNumber()}
-          />
-        ) : (
-          <span>--</span>
-        )}
-      </div>
-      <div>Submitted</div>
-      <p className="text-black">{fromNow(reward.createdAtBlockTimestamp)} </p>
-      <div>Rewarded</div>
-      <div className="text-black" color="dark.3">
-        {showRewarded ? (
-          <RewardBadge
-            amount={fromTokenAmount(contractReward[0].toString())}
-            token={token?.symbol ?? "Unknown Token"}
-            rMETRIC={contractReward[1].toNumber()}
-          />
-        ) : (
-          <span>--</span>
-        )}
-      </div>
-      <div>Status</div>
-      {hasClaimed === false ? (
-        <ClaimButton reward={reward} wallets={wallets} tokens={tokens} />
-      ) : hasClaimed === true ? (
-        <span>Claimed</span>
-      ) : (
-        <></>
-      )}
-    </Card>
-  );
-}
-
-const machine = createBlockchainTransactionStateMachine<ClaimRewardContractData>();
-function ClaimButton({
-  reward,
-  wallets,
-  tokens,
-}: {
-  reward: SubmissionWithServiceRequest;
-  wallets: Wallet[];
-  tokens: Token[];
-}) {
-  const [confirmedModalOpen, setConfirmedModalOpen] = useState(false);
-  const [successModalOpen, setSuccessModalOpen] = useState(false);
-
-  const tokenAbrev = toTokenAbbreviation(reward.sr.configuration.pToken ?? "", tokens);
-  const networkName = toNetworkName(reward.sr.configuration.pToken ?? "", tokens);
-  const wallet = wallets.find((w) => w.networkName === networkName);
-
-  const [state, send] = useMachine(
-    machine.withContext({
-      contractData: {
-        laborMarketAddress: reward.laborMarketAddress,
-        payoutAddress: wallet?.address ?? "",
-        submissionId: reward.id,
-      },
-    }),
-    {
-      actions: {
-        ...defaultNotifyTransactionActions,
-      },
-    }
-  );
-  invariant(state.context.contractData, "Contract data should be defined");
-
-  const onWriteSuccess = (result: SendTransactionResult) => {
-    transitionModal();
-    send({ type: "SUBMIT_TRANSACTION", transactionHash: result.hash, transactionPromise: result.wait(1) });
-  };
-
-  function transitionModal() {
-    closeConfirmedModal();
-    openSuccessModal();
-  }
-
-  function openConfirmedModal() {
-    setConfirmedModalOpen(true);
-  }
-
-  function closeConfirmedModal() {
-    setConfirmedModalOpen(false);
-  }
-
-  function openSuccessModal() {
-    setSuccessModalOpen(true);
-  }
-
-  function closeSuccessModal() {
-    setSuccessModalOpen(false);
-  }
-
-  const [error, setError] = useState<EthersError>();
-  const onPrepareTransactionError = (error: EthersError) => {
-    setError(error);
-  };
-
-  return (
-    <>
-      <Button onClick={openConfirmedModal}>Claim</Button>
-      <Modal isOpen={confirmedModalOpen} onClose={closeConfirmedModal} title="Claim your reward!">
-        {wallet ? (
-          <div className="space-y-5 mt-5">
-            <div className="space-y-2">
-              <div className="flex items-center">
-                <img alt="" src="/img/trophy.svg" className="h-8 w-8" />
-                <p className="text-yellow-700 text-2xl ml-2">{`todo ${tokenAbrev}`}</p>
-              </div>
-              <div className="flex border-solid border rounded-md border-trueGray-200">
-                <p className="text-sm font-semiboldborder-solid border-0 border-r border-trueGray-200 p-3">
-                  {networkName}
-                </p>
-                <div className="flex items-center p-3">
-                  <CheckCircleIcon className="mr-1 text-lime-500 h-5 w-5" />
-                  <p className="text-sm text-gray-600">
-                    {wallet?.address && wallet?.address.length < 30
-                      ? wallet?.address
-                      : `${wallet?.address.slice(0, 16)}...${wallet?.address.slice(-14)}`}
-                  </p>
-                </div>
-              </div>
-              <p className="text-xs">
-                To change or update this address head to{" "}
-                <Link to="/app/rewards/addresses" className="text-blue-600">
-                  Payout Addresses
-                </Link>
-              </p>
-            </div>
-            {error && <RPCError error={error} />}
-            <div className="flex gap-2 justify-end">
-              <Button variant="cancel" onClick={closeConfirmedModal}>
-                Cancel
-              </Button>
-              {!error && (
-                <ClaimRewardWeb3Button
-                  data={state.context.contractData}
-                  onWriteSuccess={onWriteSuccess}
-                  onPrepareTransactionError={onPrepareTransactionError}
-                />
-              )}
-            </div>
-          </div>
-        ) : (
-          <p className="my-5">
-            No address found for <b>{networkName}</b>. To add an address head to{" "}
-            <Link to="/app/rewards/addresses" className="text-blue-600">
-              Payout Addresses
-            </Link>
-          </p>
-        )}
-      </Modal>
-      <Modal isOpen={successModalOpen} onClose={closeSuccessModal}>
-        <div className="mx-auto space-y-7">
-          <img src="/img/check-circle.svg" alt="" className="mx-auto" />
-          <div className="space-y-2">
-            <h1 className="text-center text-2xl font-semibold">Claim proccessing</h1>
-            <p className="text-gray-500 text-center text-md">
-              {"This transaction could take up to {x amount of time}. Please check back in a bit."}
-            </p>
-            <p className="text-gray-500 text-center text-sm">{"If there are any issues please reach out on Discord"}</p>
-          </div>
-          <div className="flex gap-2 w-full">
-            <Button variant="cancel" fullWidth onClick={closeSuccessModal}>
-              Cancel
-            </Button>
-            <Button fullWidth onClick={closeSuccessModal}>
-              Got it
-            </Button>
-          </div>
-        </div>
-      </Modal>
     </>
   );
 }
